requirements.txt
# pip install -r requirements.txt

flask
beautifulsoup4
requests
lxml
flask-sqlalchemy
pymysql
cryptography
python-dotenv